backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
pymupdf
python-docx
spacy
firebase-admin
pydantic
python-dotenv
requests
fpdf2

